+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Version 0.8 - 2018-01-15
+## Version 0.9 - 2018-02-28
+
+- Added input control mechanisms for Apple TV
+- Added album artwork support
+
+## Version 0.8 - 2018-02-22
 
 - Added support to trigger the playback of Playlists
 - Improved error handling